working to make tui work with new parser
